+++ CHANGELOG.md
@@ -5,11 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Unreleased
+### Additions
+- New `--meta` command line option for specifying what type of metadata should be generated:
+  - `poetry` is the default value, same behavior you're used to in previous versions
+  - `setup` will generate a pyproject.toml with no Poetry information, and instead create a `setup.py` with the 
+    project info.
+  - `none` will not create a project folder at all, only the inner package folder (which won't be inner anymore)
+
+
 ## 0.7.3 - 2020-12-21
 ### Fixes
 - Spacing and extra returns for Union types of `additionalProperties` (#266 & #268). Thanks @joshzana & @packyg!
 - Title of inline schemas will no longer be missing characters (#271 & #274). Thanks @kalzoo!
 - Handling of nulls (Nones) when parsing or constructing dates (#267). Thanks @fyhertz! 
+
 
 ## 0.7.2 - 2020-12-08
 ### Fixes